+++ packages/block-library/src/missing/index.js
@@ -2,13 +2,13 @@
  * WordPress dependencies
  */
 import { __ } from '@wordpress/i18n';
-import { RawHTML } from '@wordpress/element';
 
 /**
  * Internal dependencies
  */
 import edit from './edit';
 import metadata from './block.json';
+import save from './save';
 
 const { name } = metadata;
 
@@ -18,7 +18,6 @@ export const settings = {
 	name,
 	title: __( 'Unrecognized Block' ),
 	description: __( 'Your site doesn’t include support for this block.' ),
-
 	supports: {
 		className: false,
 		customClassName: false,
@@ -26,10 +25,6 @@ export const settings = {
 		html: false,
 		reusable: false,
 	},
-
 	edit,
-	save( { attributes } ) {
-		// Preserve the missing block's content.
-		return <RawHTML>{ attributes.originalContent }</RawHTML>;
-	},
+	save,
 };
