+++ packages/block-library/src/missing/save.js
@@ -1,0 +1,9 @@
+/**
+ * WordPress dependencies
+ */
+import { RawHTML } from '@wordpress/element';
+
+export default function save( { attributes } ) {
+	// Preserve the missing block's content.
+	return <RawHTML>{ attributes.originalContent }</RawHTML>;
+}
