+++ packages/block-library/src/more/index.js
@@ -1,14 +1,7 @@
 /**
- * External dependencies
- */
-import { compact } from 'lodash';
-
-/**
  * WordPress dependencies
  */
 import { __, _x } from '@wordpress/i18n';
-import { RawHTML } from '@wordpress/element';
-import { createBlock } from '@wordpress/blocks';
 
 /**
  * Internal dependencies
@@ -16,6 +9,8 @@ import { createBlock } from '@wordpress/blocks';
 import edit from './edit';
 import icon from './icon';
 import metadata from './block.json';
+import save from './save';
+import transforms from './transforms';
 
 const { name } = metadata;
 
@@ -23,60 +18,15 @@ export { metadata, name };
 
 export const settings = {
 	title: _x( 'More', 'block name' ),
-
 	description: __( 'Content before this block will be shown in the excerpt on your archives page.' ),
-
 	icon,
-
 	supports: {
 		customClassName: false,
 		className: false,
 		html: false,
 		multiple: false,
 	},
-
-	transforms: {
-		from: [
-			{
-				type: 'raw',
-				schema: {
-					'wp-block': { attributes: [ 'data-block' ] },
-				},
-				isMatch: ( node ) => node.dataset && node.dataset.block === 'core/more',
-				transform( node ) {
-					const { customText, noTeaser } = node.dataset;
-					const attrs = {};
-					// Don't copy unless defined and not an empty string
-					if ( customText ) {
-						attrs.customText = customText;
-					}
-					// Special handling for boolean
-					if ( noTeaser === '' ) {
-						attrs.noTeaser = true;
-					}
-					return createBlock( 'core/more', attrs );
-				},
-			},
-		],
-	},
-
+	transforms,
 	edit,
-
-	save( { attributes } ) {
-		const { customText, noTeaser } = attributes;
-
-		const moreTag = customText ?
-			`<!--more ${ customText }-->` :
-			'<!--more-->';
-
-		const noTeaserTag = noTeaser ?
-			'<!--noteaser-->' :
-			'';
-
-		return (
-			<RawHTML>
-				{ compact( [ moreTag, noTeaserTag ] ).join( '\n' ) }
-			</RawHTML>
-		);
-	},
+	save,
 };
