+++ packages/block-library/src/nextpage/index.js
@@ -2,8 +2,6 @@
  * WordPress dependencies
  */
 import { __ } from '@wordpress/i18n';
-import { RawHTML } from '@wordpress/element';
-import { createBlock } from '@wordpress/blocks';
 
 /**
  * Internal dependencies
@@ -11,6 +9,8 @@ import { createBlock } from '@wordpress/blocks';
 import edit from './edit';
 import icon from './icon';
 import metadata from './block.json';
+import save from './save';
+import transforms from './transforms';
 
 const { name } = metadata;
 
@@ -18,41 +18,15 @@ export { metadata, name };
 
 export const settings = {
 	title: __( 'Page Break' ),
-
 	description: __( 'Separate your content into a multi-page experience.' ),
-
 	icon,
-
 	keywords: [ __( 'next page' ), __( 'pagination' ) ],
-
 	supports: {
 		customClassName: false,
 		className: false,
 		html: false,
 	},
-
-	transforms: {
-		from: [
-			{
-				type: 'raw',
-				schema: {
-					'wp-block': { attributes: [ 'data-block' ] },
-				},
-				isMatch: ( node ) => node.dataset && node.dataset.block === 'core/nextpage',
-				transform() {
-					return createBlock( 'core/nextpage', {} );
-				},
-			},
-		],
-	},
-
+	transforms,
 	edit,
-
-	save() {
-		return (
-			<RawHTML>
-				{ '<!--nextpage-->' }
-			</RawHTML>
-		);
-	},
+	save,
 };
