+++ packages/block-library/src/nextpage/save.js
@@ -1,0 +1,12 @@
+/**
+ * WordPress dependencies
+ */
+import { RawHTML } from '@wordpress/element';
+
+export default function save() {
+	return (
+		<RawHTML>
+			{ '<!--nextpage-->' }
+		</RawHTML>
+	);
+}
