+++ packages/block-library/src/pullquote/shared.js
@@ -1,0 +1,2 @@
+export const SOLID_COLOR_STYLE_NAME = 'solid-color';
+export const SOLID_COLOR_CLASS = `is-style-${ SOLID_COLOR_STYLE_NAME }`;
