+++ packages/block-library/src/separator/index.js
@@ -2,7 +2,6 @@
  * WordPress dependencies
  */
 import { __ } from '@wordpress/i18n';
-import { createBlock } from '@wordpress/blocks';
 
 /**
  * Internal dependencies
@@ -10,6 +9,8 @@ import { createBlock } from '@wordpress/blocks';
 import edit from './edit';
 import icon from './icon';
 import metadata from './block.json';
+import save from './save';
+import transforms from './transforms';
 
 const { name } = metadata;
 
@@ -17,39 +18,15 @@ export { metadata, name };
 
 export const settings = {
 	title: __( 'Separator' ),
-
 	description: __( 'Create a break between ideas or sections with a horizontal separator.' ),
-
 	icon,
-
 	keywords: [ __( 'horizontal-line' ), 'hr', __( 'divider' ) ],
-
 	styles: [
 		{ name: 'default', label: __( 'Default' ), isDefault: true },
 		{ name: 'wide', label: __( 'Wide Line' ) },
 		{ name: 'dots', label: __( 'Dots' ) },
 	],
-
-	transforms: {
-		from: [
-			{
-				type: 'enter',
-				regExp: /^-{3,}$/,
-				transform: () => createBlock( 'core/separator' ),
-			},
-			{
-				type: 'raw',
-				selector: 'hr',
-				schema: {
-					hr: {},
-				},
-			},
-		],
-	},
-
+	transforms,
 	edit,
-
-	save() {
-		return <hr />;
-	},
+	save,
 };
