+++ packages/block-library/src/separator/save.js
@@ -1,0 +1,3 @@
+export default function save() {
+	return <hr />;
+}
