+++ packages/block-library/src/separator/transforms.js
@@ -1,0 +1,23 @@
+/**
+ * WordPress dependencies
+ */
+import { createBlock } from '@wordpress/blocks';
+
+const transforms = {
+	from: [
+		{
+			type: 'enter',
+			regExp: /^-{3,}$/,
+			transform: () => createBlock( 'core/separator' ),
+		},
+		{
+			type: 'raw',
+			selector: 'hr',
+			schema: {
+				hr: {},
+			},
+		},
+	],
+};
+
+export default transforms;
