+++ packages/block-library/src/shortcode/save.js
@@ -1,0 +1,8 @@
+/**
+ * WordPress dependencies
+ */
+import { RawHTML } from '@wordpress/element';
+
+export default function save( { attributes } ) {
+	return <RawHTML>{ attributes.text }</RawHTML>;
+}
