+++ packages/block-library/src/spacer/index.js
@@ -9,6 +9,7 @@ import { __ } from '@wordpress/i18n';
 import edit from './edit';
 import icon from './icon';
 import metadata from './block.json';
+import save from './save';
 
 const { name } = metadata;
 
@@ -16,14 +17,8 @@ export { metadata, name };
 
 export const settings = {
 	title: __( 'Spacer' ),
-
 	description: __( 'Add white space between blocks and customize its height.' ),
-
 	icon,
-
 	edit,
-
-	save( { attributes } ) {
-		return <div style={ { height: attributes.height } } aria-hidden />;
-	},
+	save,
 };
