+++ packages/block-library/src/spacer/save.js
@@ -1,0 +1,3 @@
+export default function save( { attributes } ) {
+	return <div style={ { height: attributes.height } } aria-hidden />;
+}
